Update EratosthenesSieve.cpp
int to ll

+++ src/5. Maths/EratosthenesSieve.cpp
@@ -8,7 +8,7 @@ struct EratosthenesSieve {
     for (ll i = 2; i*i <= n; i++) {
       if (isPrime[i]) {
         primes.push_back(i);
-        for (int j = i*i; j <= n; j += i)
+        for (ll j = i*i; j <= n; j += i)
           isPrime[j] = false;
       }
     }
